Refactor context
