+++ point-of-sale/src/client/components/contexts/ThemeProvider.tsx
@@ -1,11 +1,17 @@
-import React, { FC, ReactNode, useEffect, useState } from 'react';
-import { Theme, ThemeContext } from '../../hooks/useTheme';
+import React, { FC, ReactNode, useEffect, useState, createContext } from 'react';
 
-export interface ThemeProviderProps {
-    children: ReactNode;
+export type Theme = 'light' | 'dark';
+
+export interface ThemeContextState {
+    theme: Theme;
+    setTheme(theme: Theme): void;
 }
 
-export const ThemeProvider: FC<ThemeProviderProps> = ({ children }) => {
+export const ThemeContext = createContext<ThemeContextState>({} as ThemeContextState);
+
+export const ThemeProvider: FC<{
+    children: ReactNode;
+}> = ({ children }) => {
     const query = '(prefers-color-scheme: light)';
     const [theme, setTheme] = useState<Theme>(() =>
         typeof window !== 'undefined' && window.matchMedia && window.matchMedia(query).matches ? 'light' : 'dark'
