+++ point-of-sale/src/client/hooks/useTheme.ts
@@ -1,13 +1,5 @@
-import { createContext, useContext } from 'react';
-
-export type Theme = 'light' | 'dark';
-
-export interface ThemeContextState {
-    theme: Theme;
-    setTheme(theme: Theme): void;
-}
-
-export const ThemeContext = createContext<ThemeContextState>({} as ThemeContextState);
+import { useContext } from 'react';
+import { ThemeContext, ThemeContextState } from '../components/contexts/ThemeProvider';
 
 export function useTheme(): ThemeContextState {
     return useContext(ThemeContext);
